requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.65.0
requests>=2.28.0
python-dotenv>=0.19.0
diffusers>=0.24.0
transformers>=4.36.0
accelerate>=0.25.0
psutil>=5.9.0
gradio>=4.0.0 